pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<4"
pydantic-settings = ">=2.12.0,<3.0.0"
sqlalchemy = "^2.0.46"
toml = "^0.10.2"
structlog = "^25.5.0"
orjson = "^3.11.7"
inflection = "^0.5.1"
pandas = "^3.0.1"
openpyxl = "^3.1.5"
matplotlib = "^3.10.8"
seaborn = "^0.13.2"
plotly = "^6.5.2"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-asyncio = "^0.23.0"
pytest-mock = "^3.15.1"
pytest-cov = "^7.0.0"
pytest-xprocess = "^0.22.0"
httpx = "^0.28.1"
mypy = ">=1.15.0"
ruff = ">=0.4.8"

[tool.ruff]
src = ["job_finder"]
target-version = "py311"
line-length = 79
lint.select = ["E", "W", "F", "I", "B", "C4", "UP", "RUF"]
exclude = ["__pycache__/", ".git/", "*.pyc"]
format.indent-style = "space"
format.skip-magic-trailing-comma = false
format.quote-style = "double"
preview = true

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.11"
warn_return_any = false
disable_error_code = ["import-untyped"]

[[tool.mypy.overrides]]
module = "tests.*"
disable_error_code = [
    "private-usage",
]

[[tool.pyright.overrides]]
module = "tests.*"
reportPrivateUsage = "none"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--tb=short",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=xml",
    "--cov-report=html",
    "--asyncio-mode=auto"
]
markers = [
    "unit: unit tests",
    "integration: integration tests",
    "slow: slow running tests"
]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/__pycache__/*",
    "*/tests/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "pass"
]